move VersionInfo out of a configuration header
