+++ src/version.cpp
@@ -24,6 +24,7 @@
 
 #include "gitinfo.h"
 #include "version.h"
+#include "basics.h"
 
 //==========================================================================
 //
